Добавил запросы для номенклатуры в инве

+++ union-api-client_v2/src/main/kotlin/org/openapitools/client/models/InventoryNomenclatureRecordDtoV2.kt
@@ -95,7 +95,7 @@ data class InventoryNomenclatureRecordDtoV2(
     val consignment: String? = null,
 
     @Json(name = "bookkeepingInvoice")
-    val bookKeepingInvoice: Long? = null,
+    val bookKeepingInvoice: BigDecimal? = null,
 
     @Json(name = "price")
     val price: String? = null,
