+++ union_sync_impl/src/main/java/com/example/union_sync_impl/entity/InventoryNomenclatureRecordUpdate.kt
@@ -1,6 +1,7 @@
 package com.example.union_sync_impl.entity
 
 import androidx.room.Entity
+import java.math.BigDecimal
 
 @Entity
 data class InventoryNomenclatureRecordUpdate(
@@ -12,7 +13,7 @@ data class InventoryNomenclatureRecordUpdate(
     val expectedCount: Long?,
     val actualCount: Long?,
     val consignment: String?,
-    val bookKeepingInvoice: Long?,
+    val bookKeepingInvoice: BigDecimal?,
     val price: String?,
     val cancel: Boolean?
 )
